WFLY-10231 Stabilize PassivationTestCase by splitting passivating tests into separate deployments.
