+++ testsuite/integration/basic/src/test/java/org/jboss/as/test/integration/ejb/stateful/passivation/TestPassivationRemote.java
@@ -50,6 +50,8 @@ public interface TestPassivationRemote extends AutoCloseable {
 
     void addEntity(int id, String name);
 
+    void removeEntity(int id);
+
     Employee getSuperEmployee();
 
     /**
